fix added for foreign error on shop entity

+++ src/database/2016_07_03_000000_create_shops_table.php
@@ -16,7 +16,7 @@ class CreateShopsTable extends Migration
             $table->increments('id');
             $table->timestamps();
 
-            $table->integer('user_id')->unsigned();
+            $table->unsignedBigInteger('user_id')->unsigned();
             $table->foreign('user_id')
                   ->references('id')
                   ->on('users');
